requirements.txt
tweepy
ffmpeg-python
Pillow
